This was fun lol
This was more experimental for me, I followed a tutorial that gave the structure of the code.  However, the video logged in through Facebook, I log in through Gmail.  The mechanism for swiping and getting rid of any pop-ups is different as well, the script uses the send_keys method.  The tutorial set up a virtual environment, as well as an interactive environment through bash was new to me prior to this tutorial.